fix(exec): address exact-head CI failures
